targets: add new target to generate tnlists per part.
For this the templates have to iterate over a parameter explicitely given to them.
This is a good thing because now we can do advanced filtering before filling the list,
like making a list of only all the minors or only participants of the Mörderspiel, etc.

Also add type annotation for event parameter, to please PyCharm.