targets: Allow tnlist targets to specify participants, parts and or tracks. See #2
If parts is not given, it will be constructed from the event directly.
The others are constructed from the event using the parts, if they are not given, via two new get_participants and get_tracks methods.

EVENT.parts is explicitely used when determining, whether or not the part/track shortname is needed to
differentiate a lodgement/course column, even if only the column for one part/track is printed.